==> 15_tidymodels_lin.Rmd <==
# Using tidymodels to fit linear regressions

Rather than create new chapters for `tidymodels` approaches, I will instead point you to the Class Slides, and to the relevant materials in *Tidy Modeling with R* by Max Kuhn and Julia Silge.

That text is available at https://www.tmwr.org/.
