==> 17_tidymodels_next.Rmd <==
# Using tidymodels approaches: Next Steps

Rather than create new chapters for `tidymodels` approaches, I will instead point you to the relevant materials in *Tidy Modeling with R* by Max Kuhn and Julia Silge.

That text is available at https://www.tmwr.org/.
